chore: move jwt token provider package
